rename lambda to ir
