+++ src/core/blueprint/operators/call.rs
@@ -3,7 +3,7 @@ use serde_json::Value;
 use crate::core::blueprint::*;
 use crate::core::config;
 use crate::core::config::{Field, GraphQLOperationType};
-use crate::core::lambda::Expression;
+use crate::core::ir::Expression;
 use crate::core::try_fold::TryFold;
 use crate::core::valid::{Valid, ValidationError, Validator};
