+++ src/core/blueprint/operators/js.rs
@@ -1,7 +1,7 @@
 use crate::core::blueprint::FieldDefinition;
 use crate::core::config;
 use crate::core::config::{ConfigModule, Field};
-use crate::core::lambda::{Expression, IO};
+use crate::core::ir::{Expression, IO};
 use crate::core::try_fold::TryFold;
 use crate::core::valid::{Valid, Validator};
