+++ src/core/http/request_template.rs
@@ -11,7 +11,7 @@ use crate::core::config::Encoding;
 use crate::core::endpoint::Endpoint;
 use crate::core::has_headers::HasHeaders;
 use crate::core::helpers::headers::MustacheHeaders;
-use crate::core::lambda::CacheKey;
+use crate::core::ir::CacheKey;
 use crate::core::mustache::Mustache;
 use crate::core::path::PathString;
 
@@ -736,7 +736,7 @@ mod tests {
 
         use crate::core::http::request_template::tests::Context;
         use crate::core::http::RequestTemplate;
-        use crate::core::lambda::CacheKey;
+        use crate::core::ir::CacheKey;
         use crate::core::mustache::Mustache;
 
         fn assert_no_duplicate<const N: usize>(arr: [Option<u64>; N]) {
