+++ src/core/ir/cache.rs
