+++ src/core/ir/eval.rs
