+++ src/core/ir/resolver_context_like.rs
