+++ tailcall-query-plan/src/resolver.rs
@@ -5,7 +5,7 @@ use anyhow::Result;
 use async_graphql::parser::types::{Field, Selection, SelectionSet};
 use async_graphql::{Positioned, Value};
 use indenter::indented;
-use tailcall::core::lambda::{Eval, EvaluationContext, Expression, ResolverContextLike};
+use tailcall::core::ir::{Eval, EvaluationContext, Expression, ResolverContextLike};
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct Id(pub usize);
